+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     * Toggling visibility / lock of a layer by clicking on its icon now adds an undo/redo action
     * Fixed LottieFiles import
     * Fixed dropping file as object
+    * Fixed closing compositions from the tab bar
 
 ## 0.5.4